# server/openapi.yaml
openapi: 3.0.3
info:
  title: CodeWizards API
  version: "1.0.0"
  description: >
    Backend API for CodeWizards, a student tech-club platform (projects, events,
    mentorship, doubts, blogs, coding-profile tracking, gamified leaderboard, and
    club administration). All endpoints are prefixed with /api/v1 except the root
    health check.
servers:
  - url: http://localhost:5000
    description: Local development
  - url: /
    description: >
      Production — the deployed base URL is read at runtime from the
      RENDER_EXTERNAL_URL environment variable (Render.com deployment); no fixed
      production hostname is hard-coded in the app.
...
security: []
paths:
...
  /api/v1/projects:
    get:
      tags: [Projects]
      summary: List projects
      security: []
      parameters:
        - $ref: "#/components/parameters/FeaturedFilter"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: List of projects (flat array, or paginated if page+limit are both sent)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectListResponse" }
    post:
      tags: [Projects]
      summary: Create a project (admin only)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: "#/components/schemas/Project" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Not authorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Access denied (not an admin)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /api/v1/events:
    get:
      tags: [Events]
      summary: List events (with computed status and the caller's registration info)
      description: >
        Auth is optional (optionalProtect). If a valid bearer token is sent, or a
        studentId query param, each event's `registration` block reflects that
        user's registration state.
      security: []
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [upcoming, completed] }
        - name: featured
          in: query
          schema: { type: string, enum: ["true"] }
        - name: studentId
          in: query
          schema: { type: string }
          description: Optional — used instead of the bearer token's user id to compute registration status
      responses:
        "200":
          description: List of events with registration info
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/EventWithRegistration" }
    post:
      tags: [Events]
      summary: Create an event (admin only)
      description: status is auto-derived from the supplied date (past date => completed).
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/EventInputForm" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: "#/components/schemas/Event" }
        "400":
          description: Validation error / bad image upload
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Not authorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Access denied (not an admin)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /api/v1/timeline:
    get:
      tags: [Timeline]
      summary: List timeline milestones (oldest first)
      security: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: List of milestones
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TimelineListResponse" }
    post:
      tags: [Timeline]
      summary: Create a milestone (admin only)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TimelineInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: "#/components/schemas/Timeline" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Not authorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Access denied (not an admin)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /api/v1/gallery:
    get:
      tags: [Gallery]
      summary: List gallery items
      security: []
      parameters:
        - name: category
          in: query
          schema: { type: string, enum: [event, poster, team, other] }
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: List of gallery items
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GalleryListResponse" }
    post:
      tags: [Gallery]
      summary: Create a gallery item with up to 10 images (admin only)
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title: { type: string }
                category: { type: string, enum: [event, poster, team, other] }
                eventRef: { type: string }
                images:
                  type: array
                  items: { type: string, format: binary }
                  description: Up to 10 image files (field name "images")
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: "#/components/schemas/Gallery" }
        "400":
          description: Validation error / bad image upload
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Not authorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Access denied (not an admin)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /api/v1/announcements:
    get:
      tags: [Announcements]
      summary: List announcements (newest first)
      security: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: List of announcements
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AnnouncementListResponse" }
    post:
      tags: [Announcements]
      summary: Create an announcement (admin only)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnnouncementInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: "#/components/schemas/Announcement" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Not authorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Access denied (not an admin)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /api/v1/users:
    get:
      tags: [Users]
      summary: List users (email/phone/whatsapp/discord/contactPreferences omitted)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: role
          in: query
          schema: { type: string, enum: [student, senior, alumni, admin] }
        - name: domain
          in: query
          schema: { type: string }
        - name: isMentor
          in: query
          schema: { type: string, enum: ["true", "false"] }
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserListResponse" }
        "401":
          description: Not authorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    post:
      tags: [Users]
      summary: Create a user (admin only)
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/UserCreateForm" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: "#/components/schemas/User" }
        "400":
          description: Validation error, or email already registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Not authorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Access denied (not an admin)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /api/v1/resources:
    get:
      tags: [Resources]
      summary: List learning resources
      security: []
      parameters:
        - name: category
          in: query
          schema: { type: string, enum: [PDF, GitHub, YouTube, Docs, Other] }
        - name: domain
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: List of resources
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResourceListResponse" }
    post:
      tags: [Resources]
      summary: Create a resource (admin only)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResourceInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: "#/components/schemas/Resource" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Not authorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Access denied (not an admin)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /api/v1/opportunities:
    get:
      tags: [Opportunities]
      summary: List active opportunities
      security: []
      parameters:
        - name: type
          in: query
          schema: { type: string, enum: [internship, job, freelance, open_source] }
        - name: domain
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: List of opportunities (postedBy populated)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OpportunityListResponse" }
    post:
      tags: [Opportunities]
      summary: Post an opportunity (senior, alumni, or admin)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OpportunityInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: "#/components/schemas/Opportunity" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Not authorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Access denied (role not senior/alumni/admin)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /api/v1/doubts:
    get:
      tags: [Doubts]
      summary: List doubts
      security: []
      parameters:
        - name: domain
          in: query
          schema: { type: string }
        - name: resolved
          in: query
          schema: { type: string, enum: ["true", "false"] }
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: List of doubts (author and replies.author populated)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DoubtListResponse" }
    post:
      tags: [Doubts]
      summary: Post a new doubt
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, body]
              properties:
                title: { type: string }
                body: { type: string }
                domain: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: "#/components/schemas/Doubt" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Not authorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /api/v1/coding/sync:
    post:
      tags: [Coding]
      summary: Re-sync all three platforms using the handles already stored on the current user
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Synced
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: "#/components/schemas/CodingProfile" }
                  meta:
                    type: object
                    properties:
                      results:
                        type: array
                        items: { $ref: "#/components/schemas/CodingSyncResult" }
                      cooldownRemainingMs: { type: integer, example: 0 }
                      user: { $ref: "#/components/schemas/User" }
        "400":
          description: Sync failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Not authorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429":
          description: Sync cooldown still active
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CooldownResponse" }
...
  /api/v1/leaderboard:
    get:
      tags: [Leaderboard]
      summary: Get the points-based student leaderboard
      security: [{ bearerAuth: [] }]
      parameters:
        - name: period
          in: query
          schema: { type: string, enum: [all, month], default: all }
      responses:
        "200":
          description: Ranked list of students (highest points first)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/LeaderboardEntry" }
        "401":
          description: Not authorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ---------------------------------------------------------------- Point Rules
  /api/v1/point-rules:
    get:
      tags: [Point Rules]
      summary: List all point rules (admin only)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Point rules
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/PointRule" }
        "401":
          description: Not authorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Access denied (not an admin)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /api/v1/blogs:
    get:
      tags: [Blogs]
      summary: List published blog posts
      security: []
      parameters:
        - name: tag
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: List of blogs (author populated)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BlogListResponse" }
    post:
      tags: [Blogs]
      summary: Create a blog post (authenticated members)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BlogInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: "#/components/schemas/Blog" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Not authorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /api/v1/team:
    get:
      tags: [Team]
      summary: List team members
      security: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: List of team members (sorted by teamYear desc, order asc)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeamMemberListResponse" }
    post:
      tags: [Team]
      summary: Add a team member (admin only)
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/TeamMemberForm" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: "#/components/schemas/TeamMember" }
        "400":
          description: Validation error / bad image upload
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Not authorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Access denied (not an admin)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /api/v1/contact:
    get:
      tags: [Contact]
      summary: Get club contact info (returns sensible defaults if never set)
      security: []
      responses:
        "200":
          description: Contact info
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: "#/components/schemas/ContactInfo" }
    put:
      tags: [Contact]
      summary: Create or replace the club contact info (admin only)
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContactInfoInput" }
      responses:
        "200":
          description: Upserted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: "#/components/schemas/ContactInfo" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Not authorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Access denied (not an admin)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /api/v1/collaborations:
    get:
      tags: [Collaborations]
      summary: List partner collaborations
      security: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: List of collaborations
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CollaborationListResponse" }
    post:
      tags: [Collaborations]
      summary: Create a collaboration (admin only)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CollaborationInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: "#/components/schemas/Collaboration" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Not authorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Access denied (not an admin)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Send as `Authorization: Bearer <token>`. Tokens are issued by
        /api/v1/auth/register and /api/v1/auth/login, and are invalidated
        (all of them, across devices) by /api/v1/auth/logout or a password reset.
  parameters:
...
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1 }
      description: 1-based page number. Pagination only activates when both page and limit are sent.
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100 }
      description: Page size (server caps at 100). Pagination only activates when both page and limit are sent.
    FeaturedFilter:
      name: featured
      in: query
      schema: { type: string, enum: ["true"] }
      description: Pass "true" to return only featured projects
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean, example: false }
        message: { type: string, example: "Something went wrong. Please try again." }
...
    Project:
      type: object
      properties:
        _id: { type: string }
        title: { type: string }
        description: { type: string }
        techStack: { type: array, items: { type: string } }
        contributors: { type: array, items: { type: string } }
        postedBy: { type: string, nullable: true }
        githubUrl: { type: string }
        demoUrl: { type: string }
        imageUrl: { type: string }
        featured: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ProjectInput:
      type: object
      required: [title, description]
      properties:
        title: { type: string }
        description: { type: string }
        techStack: { type: array, items: { type: string } }
        contributors: { type: array, items: { type: string } }
        postedBy: { type: string, description: "User id credited for project_posted points" }
        githubUrl: { type: string }
        demoUrl: { type: string }
        imageUrl: { type: string }
        featured: { type: boolean }
    ProjectListResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: { type: array, items: { $ref: "#/components/schemas/Project" } }
        meta: { $ref: "#/components/schemas/PaginationMeta" }

    # ---------- Event ----------
    Event:
      type: object
      properties:
        _id: { type: string }
        title: { type: string }
        type: { type: string, enum: [workshop, hackathon, contest, seminar, other] }
        description: { type: string }
        date: { type: string, format: date-time }
        venue: { type: string }
        imageUrl: { type: string }
        registrationLink: { type: string }
        status: { type: string, enum: [upcoming, completed] }
        featured: { type: boolean }
        otpCode: { type: string, nullable: true }
        otpExpiresAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    EventWithRegistration:
      allOf:
        - { $ref: "#/components/schemas/Event" }
        - type: object
          properties:
            registration:
              type: object
              description: Computed per-request based on the caller's identity (or ?studentId).
              properties:
                isRegistered: { type: boolean }
                status: { type: string, enum: [registered, cancelled, attended], nullable: true }
                certificateHash: { type: string, nullable: true }
                attendedAt: { type: string, format: date-time, nullable: true }
                registeredCount: { type: integer }
    EventInputForm:
      type: object
      required: [title, date]
      properties:
        title: { type: string }
        type: { type: string, enum: [workshop, hackathon, contest, seminar, other] }
        description: { type: string }
        date: { type: string, format: date-time, description: "status is auto-derived from this date" }
        venue: { type: string }
        registrationLink: { type: string }
        featured: { type: string, enum: ["true", "false"] }
        image: { type: string, format: binary }
...
    AnnouncementInput:
      type: object
      required: [title, body]
      properties:
        title: { type: string }
        body: { type: string }
        important: { type: boolean }
...